requirements.txt
# QA Dashboard App - Requisitos Python
# Core dependencies
pandas>=2.0.0
numpy>=1.26.0
pypdf2>=3.0.0
pdfplumber>=0.11.0
tabula-py>=2.8.0
google-generativeai

# OCR support
pytesseract>=0.3.10
pdf2image
Pillow>=7.1.0

# Dashboard and visualization
streamlit>=1.2.0
plotly>=5.15.0

# Scheduling
schedule>=1.2.0

# Google dependencies
google-generativeai

# Optional: For better Java integration with tabula-py
jpype1>=1.4.0
